+++ CHANGELOG.md
@@ -70,7 +70,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 
 
--
+- Removed deprecated `every_n_val_epochs` parameter of ModelCheckpoint
 
 
 -